refactoring dor deplouyment. creating a new package for database

+++ manage.py
@@ -3,18 +3,21 @@ import os
 import subprocess
 import sys
 
-from checkz_app import db
+from checkz_web import db, create_app
 
+from flask_script import Manager
 
-def db_create_all():
+manager = Manager(create_app)
 
-    db.create_all()
+@manager.command
+def createdb(drop_first=False):
+    """Creates the database."""
+    if drop_first:
+        db.drop_all()
 
-
-def db_drop_all():
-
-    db.drop_all()
-
-if __name__ == "__main__":
-
-    db_create_all()
+if __name__ == '__main__':
+    # if sys.argv[1] == 'test' or sys.argv[1] == 'lint':
+    #     # small hack, to ensure that Flask-Script uses the testing
+    #     # configuration if we are going to run the tests
+    #     os.environ['FLACK_CONFIG'] = 'testing'
+    manager.run()
